requirements.txt
pyspark==3.2.1
-e .
numpy
boto3
python-dotenv